refactor(style): move Color to separate color mod
